feat: get carousel order before create
