+++ apps-api/backoffice/src/modules/admin/carousel/models.ts
@@ -1,18 +1,7 @@
 import { Static, t } from 'elysia'
 
 import { CarouselNavigationType, CarouselStatusType } from '../../../../__generated__/prisma'
-
-// Core DTO for a Carousel item (aligned with DB schema)
-export const Carousel = t.Object({
-  id: t.String({ description: 'Carousel ID' }),
-  imageFilePath: t.String({ description: 'Path to the carousel image file' }),
-  status: t.Enum(CarouselStatusType, { description: 'Publish status of the carousel item' }),
-  navigation: t.Enum(CarouselNavigationType, {
-    description: 'How the app should navigate when the item is tapped',
-  }),
-  order: t.Number({ description: 'Display order (ascending)' }),
-})
-export type Carousel = Static<typeof Carousel>
+import { Carousel } from '../../../dtos/carousel'
 
 // GET /admin/carousels
 export const GetCarouselsResponse = t.Array(Carousel)
